+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- added functions `set_verbosity` and `verbosity`
+
 ## [0.1.7] - 16/09/2020
 
 - added a few tests